updated engine declaration syntax to declare required kwargs
See for details fuzzbuzz/engines/__init__.py

This has the syntax for registering an engine. It is still a little hacky
because you have to know what objects are available (to be supplied to the
kwargs by bin/fuzzbuzz). I am not sure yet how to fix that.

Additionally, I will probably simplify the registration process by making it a
decorator.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>