roughed out functions BinOps will need to have
Now that I am moving to the model of constraints be flowed instead of values
(this model hasn't been until this point under specified) value based operators
(such as set Union and arthmetic addition) will need to be constraint aware. In
particular they will need to be able to decide questions of satisfiability when
the result of their calculation has already been detirmined.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>